Upgrades network ID from u8 to u16

+++ dpc/src/traits/parameters.rs
@@ -38,7 +38,7 @@ use std::{cell::RefCell, rc::Rc};
 
 #[rustfmt::skip]
 pub trait Parameters: 'static + Sized + Clone + Debug + PartialEq + Eq + Send + Sync {
-    const NETWORK_ID: u8;
+    const NETWORK_ID: u16;
 
     const NUM_INPUT_RECORDS: usize;
     const NUM_OUTPUT_RECORDS: usize;
